requirements.txt
Flask==3.0.0
tensorflow==2.17.0
numpy==1.26.2
Pillow==10.1.0
gunicorn==21.2.0
